Related work items: #150347, #151307, #151414

+++ ut_expenses/Schema Objects/Schemas/dbo/Tables/WalletReceipts.sql
@@ -2,7 +2,7 @@
 (
 	WalletReceiptId INT NOT NULL IDENTITY(1,1) 
 	,FileExtension NVARCHAR(6) NOT NULL
-	,Status INT NOT NULL
+	,Status TINYINT NOT NULL
 	,CreatedOn DATETIME NOT NULL DEFAULT GetDate()
 	,CreatedBy INT NOT NULL
 )
